test(world): 迁入并恢复旧世界任务回归测试
