+++ server/tests/world/test_world_task_event_cleanup.py
@@ -1,10 +1,5 @@
-import sys
-from pathlib import Path
 from types import SimpleNamespace
 
-server_root = str(Path(__file__).resolve().parent.parent)
-if server_root not in sys.path:
-    sys.path.insert(0, server_root)
 
 from src.world.event_cleanup_task import ExpiredEventCleanupTask
